+++ cedar-policy-validator/src/lib.rs
@@ -135,16 +135,7 @@ impl Validator {
                     // schema if it only failed when there is a known action
                     // applied to known principal/resource entity types that are
                     // not in its `appliesTo`.
-                    .chain(
-                        self.validate_action_application(
-                            p.principal_constraint(),
-                            p.action_constraint(),
-                            p.resource_constraint(),
-                        )
-                        .map(move |note| {
-                            ValidationError::with_policy_id(p.id().clone(), None, note)
-                        }),
-                    ),
+                    .chain(self.validate_template_action_application(p)),
             )
         }
         .into_iter()
@@ -175,12 +166,8 @@ impl Validator {
         // the slot filled by the appropriate value.
         Some(
             self.validate_entity_types_in_slots(p.env())
-                .chain(self.validate_action_application(
-                    &p.principal_constraint(),
-                    p.action_constraint(),
-                    &p.resource_constraint(),
-                ))
-                .map(move |note| ValidationError::with_policy_id(p.id().clone(), None, note)),
+                .map(move |note| ValidationError::with_policy_id(p.id().clone(), None, note))
+                .chain(self.validate_linked_action_application(p)),
         )
     }
 
@@ -210,9 +197,7 @@ impl Validator {
                     ValidationErrorKind::type_error(kind),
                 )
             }),
-            warnings
-                .into_iter()
-                .map(|kind| ValidationWarning::with_policy_id(t.id().clone(), None, kind)),
+            warnings.into_iter(),
         )
     }
 }
@@ -373,6 +358,7 @@ mod test {
             r#"permit(principal == some_namespace::User::"Alice", action, resource in ?resource);"#,
         )
         .expect("Parse Error");
+        let loc = t.loc().clone();
         set.add_template(t)
             .expect("Template already present in PolicySet");
 
@@ -432,7 +418,7 @@ mod test {
         );
         let invalid_action_err = ValidationError::with_policy_id(
             id,
-            None,
+            loc.clone(),
             ValidationErrorKind::invalid_action_application(false, false),
         );
         assert!(result.validation_errors().any(|x| x == &undefined_err));
@@ -461,7 +447,7 @@ mod test {
         let id = ast::PolicyID::from_string("link3");
         let invalid_action_err = ValidationError::with_policy_id(
             id,
-            None,
+            loc.clone(),
             ValidationErrorKind::invalid_action_application(false, false),
         );
         assert!(result